requirements.txt
Flask==2.1.1
requests==2.27.1
python-dotenv==0.20.0
seaborn==0.11.2
numpy==1.22.3
pandas==1.4.2